% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_replace.R
\name{col_replace}
\alias{col_replace}
\title{Title}
\usage{
col_replace(
  df = "dataframe",
  str_pattern = "what to replace",
  str_replace = "how to replace it"
)
}
\arguments{
\item{df}{A dataframe which column names should be modified.}

\item{str_pattern}{A string with the pattern to be replaced.}

\item{str_replace}{A string with the pattern to replace with.}
}
\value{
The dataframe with the columns modified.
}
\description{
Title
}
\examples{
df <- data.frame(pre_col1 = c(1,2,3), pre_col2 = c("a", "b", "c"))
col_replace(df, "^pre_", "")
}
